+++ issue-tracking-orm/src/main/java/swt6/orm/dao/interfaces/ProjectDao.java
@@ -6,5 +6,5 @@ public interface ProjectDao {
     Project create(Project project);
     Project update(Project project);
     void delete(Project project);
-    Project findById(int id);
+    Project findById(long id);
 }
